requirements.txt
# Config
python-dotenv

# Scraping
lightnovel-crawler
beautifulsoup4
requests

# Proofreading & AI prompts (local, via LM Studio)
openai

# Voice / TTS
kokoro
soundfile

# Video stitching
moviepy>=1.0.3,<2

# Higgsfield video generation API
higgsfield-client

# Podcast publishing
internetarchive
mutagen

# YouTube / TikTok publishing
google-api-python-client
google-auth-oauthlib
